Update wish list functionality.

+++ WebAPI/Controllers/WishListController.cs
@@ -1,25 +1,34 @@
 ﻿using HM.BLL.Interfaces;
 using HM.BLL.Models;
-using HM.BLL.Services;
+using HM.DAL.Constants;
+using HM.DAL.Entities;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace HM.WebAPI.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class WishListController(IWishListService wishListService) : ControllerBase
+    public class WishListController(
+        IWishListService wishListService,
+        IStatisticsService statisticsService
+        ) : ControllerBase
     {
         /// <summary>
-        /// Retrieves the wish list for the specified user.
+        /// Allows administrators to retrieve the wish list for the specified user.
         /// </summary>
         /// <param name="userId">The ID of the user.</param>
         /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>The wish list for the specified user.</returns>
+        /// <response code="200">Returns the wish list for the specified user.</response>
+        /// <response code="401">Indicates that the user is unauthenticated.</response>
+        /// <response code="403">Indicates that the user has no permission to the action.</response>
+        /// <response code="404">Indicates that the user has no wishList.</response>
+        [Authorize(Roles = DefaultRoles.Administrator)]
         [HttpGet("{userId}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult<WishListDto>> GetWishList(int userId, CancellationToken cancellationToken)
+        public async Task<ActionResult<WishListDto>> GetWishList(string userId, CancellationToken cancellationToken)
         {
             var wishList = await wishListService.GetWishListAsync(userId, cancellationToken);
             if (wishList == null)
@@ -28,45 +37,90 @@ namespace HM.WebAPI.Controllers
             }
             return Ok(wishList);
         }
+
         /// <summary>
-        /// Adds a product to the wish list.
+        /// Allows a registered user to retrieve their with list.
         /// </summary>
-        /// <param name="userId">The ID of the user.</param>
+        /// <param name="cancellationToken">A cancellation token.</param>
+        /// <response code="200">Returns the user's wish list.</response>
+        /// <response code="401">Indicates that the user is unauthenticated.</response>
+        /// <response code="404">Indicates that the user has no wishList.</response>
+        [Authorize]
+        [HttpGet("myWishList")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public async Task<ActionResult<WishListDto>> GetMyWishList(CancellationToken cancellationToken)
+        {
+            string? userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+            var wishList = await wishListService.GetWishListAsync(userId, cancellationToken);
+            if (wishList == null)
+            {
+                return NotFound();
+            }
+            return Ok(wishList);
+        }
+
+        /// <summary>
+        /// Allows a registered user to add a product to their wish list.
+        /// </summary>
         /// <param name="productId">The ID of the product to add.</param>
         /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>An operation result indicating success or failure.</returns>
-        [HttpPost("{userId}/products/{productId}")]
+        /// <response code="200">Indicates that product was added and returns the user's wish list.</response>
+        /// <response code="400">Indicates that the product was not added to the user's wishList
+        /// and return the error message.</response>
+        /// <response code="401">Indicates that the user is unauthenticated.</response>
+        [Authorize]
+        [HttpPost("myWishList/{productId}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult<OperationResult<WishListDto>>> AddProductToWishList(int userId, int productId, CancellationToken cancellationToken)
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        public async Task<ActionResult<WishListDto>> AddProductToWishList(int productId, CancellationToken cancellationToken)
         {
+            string? userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
             var result = await wishListService.AddProductToWishListAsync(userId, productId, cancellationToken);
             if (!result.Succeeded)
             {
                 return BadRequest(result.Message);
             }
-            return Ok(result);
+            await statisticsService.AddToProductNumberWishlistAdditionsAsync(productId);
+            return Ok(result.Payload);
         }
+
         /// <summary>
-        /// Removes a product from the wish list.
+        /// Allows a registered user to remove product from their wish list.
         /// </summary>
-        /// <param name="userId">The ID of the user.</param>
         /// <param name="productId">The ID of the product to remove.</param>
         /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>An operation result indicating success or failure.</returns>
-        [HttpDelete("{userId}/products/{productId}")]
+        /// <response code="200">Indicates that product was removed and returns the user's wish list.</response>
+        /// <response code="400">Indicates that the product was not removed from the user's wishList
+        /// and return the error message.</response>
+        /// <response code="401">Indicates that the user is unauthenticated.</response>
+        [Authorize]
+        [HttpDelete("myWishList/{productId}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult<OperationResult<WishListDto>>> RemoveProductFromWishList(int userId, int productId, CancellationToken cancellationToken)
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        public async Task<ActionResult<WishListDto>> RemoveProductFromWishList(int productId, CancellationToken cancellationToken)
         {
+            string? userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
             var result = await wishListService.RemoveProductFromWishListAsync(userId, productId, cancellationToken);
             if (!result.Succeeded)
             {
                 return BadRequest(result.Message);
             }
-            return Ok(result);
+            return Ok(result.Payload);
         }
     }
 }
